Update version helpers to handle pyproject.toml
