+++ helpers/version/action.yml
@@ -5,6 +5,10 @@ inputs:
     description: The type of target to check against (core, supervisor, plugin, generic)
     required: false
     default: 'generic'
+  metadata-file:
+    description: Name of metadata file, 'setup.cfg' or 'pyproject.toml'
+    required: false
+    default: 'setup.cfg'
 outputs:
   version:
     description: The version found for the requested type
@@ -30,6 +34,8 @@ runs:
 
     - shell: bash
       id: version
+      env:
+        PYTHONPATH: ${{ github.action_path }}/../../vendor/
       run: |
         version=$(echo "${{ github.ref }}" | awk -F"/" '{print $NF}' )
 
@@ -51,7 +57,7 @@ runs:
           python3 -m pip install packaging
           python3 -m pip install .
           python3 script/version_bump.py nightly
-          version="$(python3 ${{ github.action_path }}/../read_version_cfg.py setup.cfg)"
+          version="$(python3 ${{ github.action_path }}/../read_version.py "${{ inputs.metadata-file }}")"
         fi
 
         if [[ ! -z "${{ github.event.inputs.stable }}" ]]; then
